Internal: Move site features data out of the screen to break import cycle [ED-25248]
Re-exporting CORE_FEATURE_IDS from the SiteFeatures screen made analytics
depend on React components, which closed an import/no-cycle loop and failed
Lint JS. The feature list now lives in a leaf module both sides can import.

Also waits for the install-theme request in the theme selection E2E so the
fire-and-forget install cannot race the assertion.
