+++ packages/apps/onboarding/src/analytics/core-feature-ids.ts
@@ -1,1 +1,1 @@
-export { CORE_FEATURE_IDS } from '../steps/screens/site-features';
+export { CORE_FEATURE_IDS } from '../steps/components/site-features/feature-options';
