+++ packages/apps/onboarding/src/steps/components/site-features/index.ts
@@ -1,2 +1,3 @@
-export { FeatureGrid, type FeatureOption } from './feature-grid';
+export { FeatureGrid } from './feature-grid';
+export { COOKIE_CONSENT_FEATURE_ID, CORE_FEATURE_IDS, FEATURE_OPTIONS, type FeatureOption } from './feature-options';
 export { ProPlanNotice } from './pro-plan-notice';
